Enemies spawn by pressing the N key for now. Ranger shoots arrows. No collision detection yet and the arrow is drawn poorly for now.

+++ Game/src/Util/EnemyFactory.java
@@ -7,7 +7,7 @@ import Entities.Goblin;
 public class EnemyFactory 
 {
 	
-	public static void spawnEnemy( EnemyType enemyType, int xCoord, int yCoord)
+	public static void spawnEnemy( EnemyType enemyType, float xCoord, float yCoord)
 	{
 		Enemy enemy;
 		if(enemyType == EnemyType.GOBLIN)
